Add time counting for YUT. Add unit test for hash_map. Add searching benchmark for skip_list.

+++ value.h
@@ -104,7 +104,7 @@ struct sknd {
 
 struct skls {
 	GC_HEAD;
-	unsigned short count;
+	int count;
 	unsigned short lv;
 	struct sknd *head;
 };
